Split watchdog from interceptor
